feat: 채팅방 생성 및 컴포넌트 분리
